Automatically add default SSH keys
Since the upgrade to OSX Sierra, tmux no longer seems to have access to
the OSX keychain directly, as a workaround, this additional script will
simply authenticate the default SSH keys through the OSX keychain on the
main terminal process, which does have access to the keychain. Once
authenticated, the sessions are carried over the tmux which means no
more passphrases required.